requirements.txt
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.14.3
billiard==4.2.1
blinker==1.9.0
cachetools==6.2.2
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.41.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.2
django-celery-beat==2.7.0
django-cloudinary-storage==0.3.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.16.1
Faker==37.1.0
fastapi==0.121.3
gitdb==4.0.12
GitPython==3.1.45
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.1
MarkupSafe==3.0.3
msgpack==1.1.0
narwhals==2.13.0
numpy==2.3.5
openai==2.16.0
packaging==24.1
pandas==2.3.3
paypalrestsdk==1.13.3
pillow==11.0.0
prompt_toolkit==3.0.50
protobuf==6.33.2
psycopg2-binary==2.9.10
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.22
pytz==2025.2
redis==5.2.1
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
service-identity==24.2.0
setuptools==75.8.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.1
starlette==0.50.0
streamlit==1.52.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==7.2
